refactor(nodes): simplify MultiLoraLoader into MultiLoraLoader_S2V with caching and streamlined inputs
- Removed text parsing logic, SelectedLoras, LoraItemsParser, and LoraTextExtractor classes
- Dropped dependency on regex-based prompt parsing and inline LoRA extraction
- Introduced MultiLoraLoader_S2V class focused solely on applying LoRA stacks
- Added path resolution via folder_paths for relative LoRA names
- Implemented caching to avoid reloading identical LoRA files multiple times
- Added validation and error handling for LoRA strength values
- Updated category to "Script To Video Suite" for better organization
