+++ __init__.py
@@ -5,7 +5,7 @@ from .comfyui_script_to_video_suite.s2v_nodes.s2v_storyboard_node import Storybo
 from .comfyui_script_to_video_suite.s2v_nodes.s2v_prompt_gen_node import PromptGenerator
 from .comfyui_script_to_video_suite.s2v_nodes.s2v_executor_nodes import PromptUnpacker, IterativeExecutor
 from .comfyui_script_to_video_suite.s2v_nodes.s2v_auto_lora_node import AutoLoraLoader_S2V
-from .comfyui_script_to_video_suite.s2v_nodes.MultiLoraLoader import MultiLoraLoader
+from .comfyui_script_to_video_suite.s2v_nodes.s2v_multi_lora_loader_node import MultiLoraLoader_S2V
 
 
 
@@ -16,7 +16,7 @@ NODE_CLASS_MAPPINGS = {
     "PromptUnpacker_S2V": PromptUnpacker,
     "IterativeExecutor_S2V": IterativeExecutor,
     "AutoLoraLoader_S2V": AutoLoraLoader_S2V,
-    "MultiLoraLoader": MultiLoraLoader,
+    "MultiLoraLoader_S2V": MultiLoraLoader_S2V,
 }
 
 NODE_DISPLAY_NAME_MAPPINGS = {
@@ -26,7 +26,7 @@ NODE_DISPLAY_NAME_MAPPINGS = {
     "PromptUnpacker_S2V": "4. Prompt Unpacker (S2V)",
     "IterativeExecutor_S2V": "5. Iterative Executor (S2V)",
     "AutoLoraLoader_S2V": "6. Gemini Auto LoRA Loader (S2V)",
-    "MultiLoraLoader": "7. Multi LoRA Loader (S2V)",
+    "MultiLoraLoader_S2V": "7. Multi LoRA Loader (S2V)",
 }
 
 # --- A confirmation message that your package was loaded ---
